refactor: recording adapter test package + common class RecordingImporter
