+++ backend/tests/adapters/lock/__init__.py
@@ -1,0 +1,13 @@
+from .mocks import RecordingImporter
+from .registry import RecordingLock, RecordingLockRegistry, install_recording_lock_registry
+from .timeline import LockAction, LockEvent, LockTimeline
+
+__all__ = [
+    "LockAction",
+    "LockEvent",
+    "LockTimeline",
+    "RecordingImporter",
+    "RecordingLock",
+    "RecordingLockRegistry",
+    "install_recording_lock_registry",
+]
