+++ backend/tests/adapters/lock/mocks/__init__.py
@@ -1,0 +1,3 @@
+from .importer import RecordingImporter
+
+__all__ = ["RecordingImporter"]
